[Cache] move Trait directly into Cache class
