--- src/Symfony/Component/Cache/MemoryUsageCalculator.php
@@ -1,17 +1,0 @@
-<?php
-
-/*
- * This file is part of the Symfony package.
- *
- * (c) Fabien Potencier <fabien@symfony.com>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace Symfony\Component\Cache;
-
-interface MemoryUsageCalculator
-{
-    public function calculate(): int;
-}
